+++ src/main/java/main/Main.java
@@ -3,7 +3,11 @@ package main;
 import org.eclipse.jetty.server.Server;
 import org.eclipse.jetty.servlet.ServletContextHandler;
 import org.eclipse.jetty.servlet.ServletHolder;
+import org.glassfish.hk2.utilities.binding.AbstractBinder;
+import org.glassfish.jersey.server.ResourceConfig;
 import org.glassfish.jersey.servlet.ServletContainer;
+import rest.Users;
+import rest.Session;
 
 /**
  * @author iu6team
@@ -24,8 +28,20 @@ public class Main {
         final Server server = new Server(port);
         final ServletContextHandler contextHandler = new ServletContextHandler(server, "/api/", ServletContextHandler.SESSIONS);
 
-        final ServletHolder servletHolder = new ServletHolder(ServletContainer.class);
-        servletHolder.setInitParameter("javax.ws.rs.Application","main.RestApplication");
+        final Context context = new Context();
+        context.put(AccountService.class, new AccountServiceImpl());
+
+        final ResourceConfig config = new ResourceConfig(Users.class, Session.class);
+        config.register(new AbstractBinder() {
+            @Override
+            protected void configure() {
+                bind(context);
+            }
+        });
+
+        final ServletHolder servletHolder = new ServletHolder(new ServletContainer(config));
+
+        //servletHolder.setInitParameter("javax.ws.rs.Application","main.RestApplication");
 
         contextHandler.addServlet(servletHolder, "/*");
         server.start();
