+++ src/main/java/rest/Session.java
@@ -1,6 +1,9 @@
 package rest;
-import main.AccountService;
 
+import main.AccountService;
+import main.AccountServiceImpl;
+
+import javax.inject.Inject;
 import javax.inject.Singleton;
 import javax.servlet.http.HttpServletRequest;
 import javax.ws.rs.*;
@@ -17,15 +20,18 @@ import javax.ws.rs.core.Response;
 @Singleton
 @Path("/session")
 public class Session {
-    private AccountService accountService;
+    @Inject
+    private main.Context context;
+    //private AccountServiceImpl accountService;
 
-    public Session(AccountService accountService) {
-        this.accountService = accountService;
-    }
+    //public Session(AccountServiceImpl accountService) {
+    //    this.accountService = accountService;
+    //}
 
     @GET
     @Produces(MediaType.APPLICATION_JSON)
     public Response checkAuth(@Context HttpServletRequest request) {
+        final AccountService accountService = context.get(AccountService.class);
         final String sessionId = request.getSession().getId();
         if (accountService.checkAuth(sessionId)) {
             UserProfile userTemp = accountService.giveProfileFromSessionId(sessionId);
@@ -41,6 +47,7 @@ public class Session {
     @Consumes(MediaType.APPLICATION_JSON)
     @Produces(MediaType.APPLICATION_JSON)
     public Response loginUser(UserProfile user, @Context HttpHeaders headers, @Context HttpServletRequest request) {
+        final AccountService accountService = context.get(AccountService.class);
         if (accountService.isExists(user)) {
             if (user.getPassword().equals(accountService.getUser(user.getLogin()).getPassword())) {
                 final String sessionId = request.getSession().getId();
@@ -54,6 +61,7 @@ public class Session {
     @DELETE
     @Produces(MediaType.APPLICATION_JSON)
     public Response logOut(@Context HttpServletRequest request){
+        final AccountService accountService = context.get(AccountService.class);
         final String sessionId = request.getSession().getId();
         accountService.deleteSession(sessionId);
         return Response.status(Response.Status.OK).build();
